// Wapi-admin/src/types/template.ts
export interface AuthData {
  add_security_recommendation?: boolean;
  code_expiration_minutes?: number | string;
  otp_buttons?: { otp_type: string; copy_button_text?: string }[];
}

export interface TemplatePreviewBubbleProps {
  templateType: string;
  headerText: string;
  bodyText: string;
  footerText: string;
  buttons?: any[];
  fileUrl?: string | null;
  marketingType?: string;
  offerText?: string;
  productCards?: any[];
  mediaCards?: any[];
  mediaButtonTemplates?: any[];
  authData?: AuthData;
  platform?: string;
}